rearanged tests
